requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-dotenv
pydantic[email]
